% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notly.R
\name{ggplotly}
\alias{ggplotly}
\title{Turn a ggplot into a plotly object}
\usage{
ggplotly(
  p = ggplot2::last_plot(),
  width = NULL,
  height = NULL,
  tooltip = "all",
  dynamicTicks = FALSE,
  layerData = 1,
  originalData = TRUE,
  source = "A",
  ...
)
}
\arguments{
\item{p}{ggplot object}
}
\value{
Plotly object with a ggplot nested inside
}
\description{
This function takes a ggplot object and turns it into a plotly object.
The original ggplot object is retained inside of the new plotly object.
This makes this conversion reversible, allowing you to freely go back and forth betweeen ggplot and plotly.
}
