% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notly.R
\name{notly}
\alias{notly}
\title{Turn a notly back into a ggplot object}
\usage{
notly(notly_object)
}
\arguments{
\item{notly_object}{notly object}
}
\value{
ggplot object
}
\description{
This function takes a notly object (plotly with embedded ggplot) and turns it into a ggplot object.
It is the inverse function to \code{ggplotly()}, allowing you to freely go back and forth betweeen ggplot and plotly.
}
